Ensure we handle arithmetic operations on all bitnesses correctly.
We adopt a single, simple type promotion rule for arithmetic types
(integers, booleans, and floating point numbers).  When we are operating
on different types, we promote them to their common type, and perform
the operation on that type (except for a few exceptions relating to powers).

The promotion of any two arithmetic types requires us to make the result
floating point if either argument is floating point, signed if either is
signed, and to take the max bitness.

This is a departure from C++, because C++ compares signed and unsigned
integers as unsigned integers. But we need this behavior if we are going
to be consistent with python bool-to-integer comparisons which treat the
bool like the integers 0 or 1.