If branches understand 'isinstance' and 'match' and can restrict types within based on it.
In particular, we understand that if we write `if isinstance(x, T)` that within the if branch,
x is _definitely_ a 'T', which means our type inference is more precise, and also means that
we avoid some expensive runtime checks.

Similarly, we understand if 'x.matches.T' if 'x' is an Alternative means that 'x' is in fact
the type that matches.