Ensure we handle type objects more carefully. Refs #153.
We ensure that we have a typed_python.Value representation for all type
objects and free python objects (e.g. `range`) so that we can correctly
model variables that have OneOf different values of these types.

You can now request a value like Value(int), which means the 'int type object'
as a typed_python type. This means you can say something like OneOf(Value(int), Value(float))
to mean 'either the int type object or the float type object'. This makes
our one-of code more consistent.

We still have some nasty places we're asking whether an expression's
expr_type is a PythonFreeObjectWrapper and then dispatching based on that object's
typeRepresentation. It would be better to pick a simpler-to-maintain interface.