Fix a bug in type-conversion to ListOf/TupleOf. Refs #219.
Type-conversion is supposed to proceed in two passes - first, with 'explicit' set to
False, which is intended to not do any deep conversion, and then again with 'explicit'
set to True, which allows a broader set of conversions. This is intended to allow us
to have multiple overloads of function types and to get approximately correct behavior
without forcing the very first specialization to work. Similarly, it allows
OneOf(int, float) and OneOf(int, str) to both accept floats.

We had two bugs here - first, we were not obeying the rules in List/Tuple conversion.
Second, we were not clearing the exception bit when catching a 'PythonExceptionSet'
exception, which leads to corruption of the python thread state.