Calling 'str(T)' is not a conversion, it's a cast!
typed_python doesn't allow you to 'convert' an int to a string. If you attempt
to append an integer to a ListOf(str) you'll get an error. The compiler, OTOH,
was treating 'int.convert_to_type(str, explicit=True)' as an opportunity to convert
an int to a string, which is just wrong.  To support this behavior, T(x) resolves
to 'x.convert_cast(T)' which then resolves to T.convert_cast_to_self(x) if 'x'
doesn't know how to cast itself.