Speed up some class dispatch code.
In general, we should do some work to optimize this because it's
extremely slow! We do things like convert char* to std::string all over
the place, box and unbox our types, etc. The compiler optimizes all
of this away, but at the interpreter level it's important not to
be too much slower than standard python.

This change tries to cache some types for method dispatch and use
hash tables on const char* instead of std::string ordered maps.