Ensure we can handle classes that mix inheritance and lots of members. Refs #123.
We had a bug where we weren't aware that subclasses could change
the size of the 'initializer' bitfield (which keeps track of which members
are initialized). The base class might use 1 byte and the child class 2,
and so code compiled for the base class would index into the wrong spot
in the child class.

Now we encode this in the vtable, and use that during dispatch.
We could extend this to allow members from multiple base classes if we
wanted.