Fix a couple of build issues and roll forward the typed_python pin.
We had to remove pyproject.toml, because it causes us to use PEP 517-style
installation, which for some reason causes us to pass "--ignore-installed"
to 'pip install' when installing object_database. This is a problem because
we now have different versions of typed_python outstanding with completely different
header files.  It causes us to build object_database's c++ binaries against a
different version of typed_python than what is actually installed in the
target virtualenv and then everything crashes.

I removed it and the build appears to work again. If we want to re-enable
'black', we'll need to to find another way to give it the information than
using a toml file.