Rework how we flatten the structure of nested Sequence and Subscribeds.
Instead of moving cells around and changing their parents when we recalculate
the cells tree, we reorganize so that we can _compute_ what the flattened tree
ought to look like.

As part of this, cells re-compute the highest dirty node first, and then
recalculate new children recursively. This makes it possible for cells to
assume that they always can know their parent when they compute, and so
therefore they can determine whether they get 'merged in' to their parent.

Then separately we compute the 'displayable children' of a cell when we
render the cell on the wire, as well as its displayable properties (such
as whether its a flex parent).