Don't send messages for parents and also children in the same message bundle.
Right now, when a cell 'P' changes, we actually produce a rendering for all of its
children and send that over the wire to the frontend, which then re-renders all of
the components and uses maquette to jam their contents into the DOM.

If a child 'C' (say, a Sheet) is inside of a parent (say, a SubscribedSequence P),
and _both_ of them change, then if we send both C and P in the same message, the
message for P will implicitly define C, and C will be rendered twice. Because
OldSheet objects only initialize handson tables once, this causes maquette
to completely destroy the table representation.

Also as part of this change, don't send so many requests for row data, so that
the table is a (little) snappier.