Make it possible for Expanders to retain state across sessions.
As part of this change, introduce an idea of an 'identityPath' which
identifies the location in the cells tree of a given cell. The current
implementation is not very sophisticated (it uses the integer position
of the cell in its parent tree). A more sophisticated version should
do something with subscribed sequence to associate the position of the
cell with the item in the sequence.

In any case, this allows us to stash the state of the Expander in a
place that's persistent across bouncing the browser, and that allows
us to open/close subtrees and retain the state of the opener.