Fixing bug where #componentDidUpdate not triggered on Subscribed contents
This came as the result of the new "abbreviated" update message structure,
where we only pass IDs for certain children that have not changed at all.

The solution implemented here was to leverage the `.isSubscribed` property
of Subscribed components in the handler, and to force-add their contents
to the list of "updated components," who will then have #componentDidUpdate
called.

I have added tests for this and related scenarios.