Updating lifecycle handling for created and updated Components.
Previously, CellHandler kept internal lists of updated and
created components, and would push to them where needed. These
are used at the end of a cycle to call #componentDidLoad (created)
and #componentDidUpdate (updated) respectively.

It turns out we could find ourselves pushing the same component
to this list multiple times, triggering multiple lifecycle method
calls in a single cycle. This led to unexpected behavior.

Instead of changing the lists to a Set, we now have properties
on the base Component prototype that have default settings. We
collect/select for these attributes once each cycle, call the
appropriate lifecycle method, and then reset the property to false.