unipin the version of typed_python from the Pipfile
The assumption is that we have typed_python checked-out at ../typed_python
at the version we want to use. This allows projects that depend on both
typed_python and object_database to check out the versions of these
libraries they want and to build an app that uses those versions.