Allow panel to disable its border. Refs #68.
This is really just a workaround for some issues with subscribed as it's
currently implemented. The new pattern of having Subscribed not have its
own div means that when we update the root of a tree of Subscribed and SubscribedSequence
objects, we redraw a lot of the DOM, which makes interacting with the page
impossible during frequent updates. Putting borderless panels around some
of these cells mitigates the problem, altough eventually we need to fix this.