Remove some css on 'expanders' that was causing modal buttons to not show up correctly.
Modals triggered from expanders were getting a 'background-color:transparent' tag which then
caused the modal buttons to be unreadable. In general, I don't think we want expander-contents
to be treated differently than other cells, so if we want to bring this behavior back
we should do it with some kind of styling mixin that we can apply to a cell.