Changing element ids to valid values and updating tests to accommodate.
Note that all tests are passing and things are working normally.
We now have two "kinds" of ids for Cell structures in the frontend:
* Element IDs: The DOM id attribute given to the root element of a
given Cell Component structure, which has a valid prefix of 'cell-*'
* Normal Cell Ids, which are the same as Component.props.ids, and
are just numbers (or str reps of numbers), and which can be queried
in the DOM also using the 'data-cell-id' attribute.