Updating Component side messaging setup and adding Modal events
We now include an instance of the CellSocket on the CellHandler,
and enable specific message sending to the Cells backend from within
both base Component class and NewCellHandler. All Components now have
a property that references the handler, should we initialize them
that way.

Additionally, we are adding 'close' and 'accept' message events to
the Modal cell. Close will simply set the show slot to false. Accept
does whatever the action is on the first available button in the Modal,
if there are any, and then sets the slot to False, closing the modal.

Close and Accept are also mapped to the Escape and Enter keys
in a showing Modal.