Adding separate TableColumn and TableColumnSorter Cell/Components
TableColumn links to a filtering slot and knows how to construct
the DisplayLineTextBox that interacts with that slot. It also
stores a key and label for the column

TableColumnSorter links up to a single filtration slot and
knows how to interact with that slot when clicking on sort
buttons. It also builds the sort buttons for each a given
column key. Note that the filter slot is a 2-dimensional
list or tuple, with the first element as the column key
and the second as either 'ascending', 'descending', or
None. If either or both elements are None, then nothing is
being sorted at all.