Attempting NewTableHeader and CellsDemos for it, but currently buggy.
This is an attempt at creating a new kind of TableHeader that
makes use of TableColumn, TableColumnSorter, TablePaginator,
and some of the shared Slots between them.

There is odd behavior that I cannot wrap my head around, which can
be reproduced in the following cells_demos:
* table_parts.BasicNewTableHeader
* table_parts.AltBasicNewTableHeader

The first example (BasicNewTableHeader) works as expected:
filtering slots for each TableColumn are shared and, when updated,
the small Panels at the bottom correctly display for each one.

The second example fails bizzarely. None of the displays on the bottom
will update based on the filter slot for each TableColumn at all, with
one exception: the last one, which, for some reason, will update all
the slots and displays. The only different between these two examples
is that the shared slots are stored in a dictionary and/or on the
TableColumn objects themselves.

This might affect my planned design for NewTable, which would
incorporate all of these pieces together using well defined slots
and structures.