Adding TableRow and intelligent TablePage Cells/Components
TableRow - This is straighforward. It's more or less a wrapper
for <tr> elements on the front end, with some extra information
like the current index in the page it's being displayed in.

TablePage - This is the "intelligent" part of the Table. For a given
list of sorted and filtered row information, and a current page num,
total page size, and max page size, it will display only the subset
of the rows that meet the calculation.