NewTable ignoring filter slot updates for now.
The TableColumn header items are now working as expected,
but actual filtering does not yet work.

Also note that we moved to have TableRow use a list of
slots rather than a list of filter terms for the filtering.
We might want to change this back after some additional
refactoring.