Refactoring DisplayLineTextBox component and demo
This represents a near full rewrite of DLTB.
The motivation is that input elements do not dynamically
resize like other flowing DOM elements. Our solution is to
have a label div and an input element and to "swap them out"
when the component becomes active (showing input) or inactive
(showing the label div). When activating the input, we first
get the label div's current width, then set the input's width
accordingly. This gives the illusion of a dynamic input field.
It should also flow well with responsive stylings like flex.