Fixing ResizablePanel Code issue (#110)
The problem was threefold.
1. Code was too complex: it wrapped a pre in a code in a Text
component. Code no longer produces a Cell child (Text) since it
is always a string. It is now passed in as exportData/props
2. ResizablePanel panes are not set up to automatically deal
with overflows.
3. Plain Pane elements are also not setup to handle overflow,
which is correct in most cases since Panel fills the space it
has. ResizablePanel is dynamic, so Panel was only taking up
the initially rendered space, and overflowed Code was spilling
out. We solve this with a specific selector for Panes that are
the direct children of a ResizablePanel item