Fixing relative/shift calculations to get tests to pass.
The fix involved resolving a conceptual misunderstanding
of what dataOffset is all about.
In the visuals and accompanying tests, I kept assuming
(in complex cases) that the dataOffset would be equivalent
to the origin of the relativeViewFrame. But this is not
the case. The x and y values are sort of independent, and
each tells us how far off from the given side PLUS any
number of locked rows/columns on that side the data is.
The getter for the relative frames will handle any differences
after that.

The tests are currently passing and this seems to work.