Ensuring that shifts automatically update elements in PrimaryFrame
Previously, a manual call to `#updateCellContents` was required
in order to update the elements in a PrimaryFrame after a shift
occurred.

We now call this after every shift.

The reason we do this is because we want to store the relative
point values on each of the PrimaryFrame's elements, for
both introspection and later viewing by something like a
Selector. We now do just that.