Full working movement and early selection/visual implementation
"Cursor" movement in all directions and across pages is
working normally.

There is an early implementation of selection, with visuals,
going on here but it is kind of a ratchet job at the moment.

Might consider separate frame with dual-cursor implementation,
but need to think about the absolute/relative terms of the
corresponding Points