Working async DataFrame implemented in test component
Note that we are not using any buffering strategy,
so each page or shift by single row/column triggers a refetch
on all relative frames that have empty data, which in the
beginning will be all of them each time.

Consider using a buffering option that is measured in page
sizes in each direction, limited by the ends of the underlying
DataFrame, of course.