NewSheet Cell/Component data communication integration
NewSheet is now a Sheet-like Cell by definition, and can
communicate with its Component over the data channel.

We are also making use of the webcomponent's
`sheet-needs-data` event for determining when to
send data requests to the Cell backend.

Note that we are also padding these requests by
(possibly) increasing the size of frames before
they are sent to the backend.

We have updated the CellsDemo to use this new
version of the NewSheet cell.