[BUGFIX] Subsequent shiftRight and shiftDown calls working (see)
Description:
A bug was found whereby if one shifted the PrimaryFrame
to the complete right or bottom ends, then attempted to
shift a single point again in that same direction, an
error would be thrown about attempting to access data
at points outside of the DataFrame.

The cause appears to have been a miscalculation when
determining how to adjust new dataOffset values in
shiftDownBy and shiftRightBy. Instead of calculating
the "potential new" right/bottom values based on the
total number of cells in the x or y column, we were using
only primaryFrame.viewFrame, which does not include any
locked rows/columns in the calculation.

Hence the reason this bug did not manifest if there were
no locked frames along the buggy axis.

We have fixed the bug and have also included tests for this
in the shift tests file.

Note also that we cleaned up the implementation of attribute
setting and adjustment in the APSheet webcomponent file.