Fixing TablePage / TableRow Slot subscrip. boundaries
Description:
We had an issue whereby a single Table-cell Cell element
that is bound to a Slot would cause the whole TablePage to
update rather than just itself. The upshot was potentially
megabytes of unnecessary message updates coming over
the wire.

The source of the trouble was that TablePage indirectly calls
`sortsAs` on elements (through it's child TableRows) in order
to determine which rows should be filtered and sorted out. For
Subscribed cells, `sortsAs` will trigger any `Slot.get` calls
within the lambda/function that is Subscribed. Therefore if
we do not set the correct ComputingCellContext before calling
`sortsAs` on an element, we run the risk of subscribing some
parent Cell to the slot accidentally. That is what was happening
here.