[BUGFIX] td elements now get correct data attributes
Description:
After we added the inner div (for styling purposes), the
PrimaryFrame's `updateCellContents` method was adding
the data attributes to itself rather than to the parent
td element. This was throwing errors in the tests and
cursor/anchor calculations.

The solution was to add a method `innerElementAt` to
TableElementsFrame (superclass of PrimaryFrame) and to
be explicit about which element was doing what in the
updateCellContents methods.