CSS Styling and Content Handling Improvements
Description:
In order to have better styling, particularly when making selections
and adjusting the cursor position,  we have made some changes to the
element structures and the CSS stylesheet.

The issues that made this change necessary were:
* Text inside of sheet cells was not vertically centered, so looked
terrible;
* Moving the cursor around or making a selection would wobble the rows
and columns as the user adjusted the boundaries, whether by keyboard
or mouse
* Large text contents were not being displayed properly

Solution:
The solution was to add a new inner-inner <span> element inside of
the td > div.sheet-cell-inner, while giving it a
.sheet-cell-inner-content class. This allows us to use flexbox on the
div, since we cannot use flexbox on the td elements themselves (this
is a matter of how tables are implemented in the DOM). The upshot is
that text is now centered.

Additionally, we have updated the stylesheet to deal more competently
with borders and padding adjustments when changing the selection frame
and cursor movement. Now things should not "wobble".

In order to facilitate this, we've also added a convenience method to
TableElementsFrame (superclass of PrimaryFrame) that allows us to set
the text contents of the inner-inner span without first getting the
element. This method is called `setTextContentAt()`. **NOTE**: this
method will also remove any newlines from incoming text contents,
since that was mucking up styling of overflow/ellipses.