[BUGFIX] Large content cells now remain the same size when selecting
Description:

Table cells where constantly adjusting size during navigation in
situations where the content of the cell was a large amount of
text. This was due to how the text is displayed in the second th
header cell, and how its dimensions are set up.

Fix:
We calculate the correct colSpan for the the display th element and
update that whenever the number of columns in the PrimaryFrame
changes. Additionally, we set the max-width of that cell to be the
same as the passed-in colWidth. This is because colSpan is calculated
based upon individual th calculations, NOT based on one element alone.

This appears to fix the particular navigation UI bug in question.