[BUGFIX] Fixing issue where sheet resized beyond parent container
Description:
The current implementation of the resize() resulted in sheets that
exceeded the bounds of the parent elements.

The issue stems from two problems:
1. We can't know for sure the bounding sizes of any td element ahead
of time. They do not appear to respect the box-model for border-box in
a detailed way. For example, any borders, padding, margin, or other
hidden measurements that table decides in its layout algorithm cannot
be known ahead of time.
2. We were using the wrong Math functions in our decimal
calculations. We should have been rounding down.
3. We were not removing any thead measurements from the row calculations

The solution to 1 involves first seeing if a td element for the
current sheet is already attached. If so, we grab its bounding
measurements which will be fairly accurate. If not, go by the initial
rowHeight/colWidth that was specified at construction. This seems to
work well.

The other solutions were straightforward fixes.