Reimplementing multi-line header features from older commit
This commit reimplements changes made in
221904a01cc79017b88c9ffb22ac1c81a51c6953

These changes were designed to display a "drop-down overlay" from the
table header when the content of the current cursor was too large to
display in a single line. Previously, in this new sheet
implementation, we were cutting off such large text displays using ellipsis.