Updating KeyAction and adding 'all' KeyBinding option
In commit d4c3adf6a89aa1e9f847b3820fb2866a47820370 Braxton had added
and fixed some parts of the old KeyListener setup. Since this branch
includes the PR for the new KeyListener rewrite (and thus erases some
of this changes) we need to adapt these.

This commit does implements two of those changes:
1. We ensure that we are deregistering (here called pausing, see code
comments) a given Component's KeyListener inside of the
componentWillUnload callback.
2. We are adding the option to specify 'all' as the key in a
keyCmd. This will trigger the event when any of the non-modifier keys
are pressed.