Replacing header display with new independent display area
This new area will take up the maximum horizontal space available to
the ap-sheet element, regardless of how many columns the underlying
table currently displays. This allows us to display the large content
text cells in a much more readable way when the table itself only has
a couple of columns.

We use a separate div and absolutely position it to the spot where
the th:second-child would normally be. We have also removed all older
references to the old header display.

NOTE: This does not affect the cursor location header display, which
is th:first-child.