Fixing header display alignment and adding new cells demo
Description:
ap-sheet was not set to be the relative offset parent, so in cases
where a sheet was inside of a panel, the content display element was
positioning itself relative to the panel. We have fixed this issue.

This also required recalculating the width of the content display in a
better way, which we've now also implemented.

There is a new cells demo BasicNewSheetInPanel which demonstrates this.