Allow the service_manager.py to watch the docker image in AWS.
Right now if you want to change the docker image, you need a new host,
which is really slow and awkward. This change allows the docker image
to be reset on in ODB and for all connected workers to see it and pull
the new image.

It also fixes a long-standing bug where some workers would fail
to connect and then just exit the docker image. Instead, our bootstrap
script will attempt to repeatedly reconnect to the main odb server.