You can't read and write from the same OpenSSL socket object at the same time.
It's not well documented, but open ssl internals do not want you to read and write
on the same OpenSSL-wrapped socket at the same time. If you do, you'll get weird
socket errors at random times when multiple threads are running.  This is in contrast
to normal posix sockets which are perfectly happy for you to read and write to them
simultaneously.

This change rewires the message_bus to use a single thread. We'll need to do the
same thing for the main ODB thread as well.