Rewrite the css model for controlling how elements flex.
In the old model, we had 'flex-child' and 'flex-parent', which sort of
worked but had weird edge cases.

In the current model, we have a formal notion of 'space preferences' for
each kind of cell, indicating whether it wants to take up as much space
as it possibly can in each of the two dimensions. We also have a notion of
whether space is 'infinite' in each of the dimensions.

Cells use this to calculate properties 'fill-space-horizontal' and
'fill-space-vertical' which they attach to their dom elements as classes.

Cells like code editors, tables, plots, etc, are all 'space fillers'.
Cells like text, buttons, etc are not. When you stack space fillers
and non-space fillers, the obvious thing happens.

This allows most code to not need 'flex' anywhere inside of it.