SupbrocessServiceManager Logging
We had an issue with logs filling up the disk and causing a firestorm of
problems resulting in an absurd number of SAPI pulls.
SubprocessServiceManager now has a log quota and clean-up logic along
with using RotatingFileHandler for logging to files which guarantees:
- log files will be at most X bytes (default 100MB) with excess logs
  being moved to logile.1, logfile.2, ..., .k. This allows us to
  prune logs even from an active service by deleting some of its backup
  logs (the 'head' log of an active service may never be deleted)
- a predetermined maxium number of .k backups is allowed, guaranteeing
  an upperbound on an active process' log size

The cleanup logic is as follows:
- delete the fewest log files to get within our log quota (default 20G)
- start by deleting from services that have logs for at least two
  terminated instances, starting from the oldest logfile
- continue by deleting from services that have logs for at least ONE
  terminated instance, starting from the oldest logfile
- continue by deleting logfiles from the active instances, starting
  from the service whose logs are the largest.

The RotatingFileHandler is also used for the odb_server logs, instead of
logging to the terminal and relying on the caller to forward stdout (and
stderr). We don't have an instanceID to use for these logs as we do for
service logs, neither can we use the PID (via os.getpid()) because when
run through docker the service_manager.py process gets one of two PIDs
depending on boot timing, which resulted in logs getting overwitten when
moved to old/ or appended when getting the same PID as the last
instance. So instead we use the integer part of the timestamp when the
service_manager.py process is started.