Fix a bug in session states.
ComputedSlot, when it recalculates, creates a new db transaction
underneath the current context. This doesn't play nicely with
writes into odb, which happens when we set up session state.

In general, we need to do a pass rectifying all the 'slot' notions
and making sure that we have a fully consistent model of reading
and writing from ODB. Basically, we shouldn't ever write into
ODB when we are 'calculating' the system, and then we need
a separate formalism around applying side effects that arise
because of inconsistencies we observe when calculating something.