Make sure we can handle very large sheets.
This is not the most elegant way to structure the code. Basically,
the DOM can't handle us creating a div that's millions of pixels
wide, and then offsetting it negatively to show a small window of it.

Instead, we need to carefully show smaller windows of the space we're
rendering so that the widths and offsets are not "too large" (for some
roundoff error that's not totally clear).

A better model than what I have here would be to formally define
the different view regions, and calculate everything relative to their
pixel offsets, along with explicit clipping regions.

But, this seems to work so we can refine it when we find a more subtle
bug.