Formalize the cells notion of computation, slots and views.
Prior to this change, we had a sort of mish-mash process for deciding how mutability in slots
interacted with the process for calculating the cells tree and computed slots. This could lead
to subtle bugs where messages from the client trigger recalculation of slots in strange ways.

Now, the model is much clearer:

* during the calculation of the cells tree, all state is read-only. You may not modify slots
  that you didn't create during that particular invocation of a Subscribed, nor may you write to
  the ODB.
* ComputedSlots compute on demand whenever they are needed. They also are not allowed to modify state.
* State can be modified by the following:
    * message handlers (which are responding to some user event that occurred on the UI)
    * scheduled callbacks, which run in-between rendering frames and are allowed to modify both the
	ODB and also slots
    * cell 'Reactor' objects, which encapsulate the idea that some portions of a cells tree
      may need to react to the slot/odb state and make some formal state change to it.

State-modifiers are intended to be transactional: if they throw an exception, then we unwind
any changes they made to both the ODB and also to the slot state.

Our original model had a notion of slots having listeners, and modifications to slot state having
a 'reason' - the primary motivation here was to distinguish changes to slots that occurred because
of database changes vs UI modifications - that way we could distinguish between a slot containing
some ui state (say, the top visible line number of a text editor) changing because the user modified
it or because a database slot changed. This was important because if the database changed it we
wanted to fire off a message to the UI telling it to scroll.

Under the new model, the idea is to maintain a slot that represents what the UI knows as the
state (say, the top line number) and separate that from the slot that contains the db-stored
version of that top line number (say, the session state).  If the database value changes,
we update this slot state and send a message to the ui telling it to scroll. If the user
scrolls, we update the slot and also write into the db. This gives us a formal model for
the current state and an explicit set of rules for how to update them that's easier
to verify correctness for.

Note that user-reactors can easily put themselves in loops - we don't do anything to prevent
that yet. Something to consider.

Also as part of this change, clean up how the error handling mechanism in all cells works -
any cell recalculate that throws an error automatically sets an 'exception' prop that all
javascript cells must know how to render (most get it for free through ConcreteCell.js).