Fully fix the transactional model in cells.
We had a number of subtle errors in the way we were handling
the interaction between the ODB and side-effects that result from
cells calculations.

In particular, we were

    (a) not tracking how writes to the ODB/slots would dirty computed slots
    (b) being careful to walk back changes to Slot objects and to unsend
        messages when a revision conflict occurred
    (c) were not ensuring that the exact view-state was in sync with the
        state of the computed slots.

As a result, a number of inconsistencies could creep into the data model.

This change fixes that by being very careful to model a full 'view' of the
ComputedSlot state during execution of side-effectful computations.