Dont allow transactions before the minimum GC'd revision id.
We have to keep track of the most recent version id where an object
was changed. Unfortunately, we cannot keep track of all the objects that
have ever been changed because that set could grow arbitrarily large. Instead
we 'forget' about any object fields that were deleted more than N seconds ago,
where N is a gc interval usually on the order of a few minutes.

This can lead to cases where we might allow conflicting transactions to succeed
if they were based on version numbers from a long time ago. One such scenario
that we repro here is to allow a deletion of an object by two transactions
separated by the GC interval. The server cleans up the record that the object
was deleted and then the second transaction, which should conflict because it
would read the ' exists' field before writing to it, fails to conflict because
we now have forgotten about the deletion.

This change ensures that we don't allow such transactions through since
they can corrupt the state of downstream clients.