Be more careful to distinguish between reactors that had effects and those that didnt.
We update the cells display until we think it has stabilized.  Reactors in cells
calculations cause us to update our view of the most recent transactions from
the ODB. Its possible for this to trigger an infinite cycle where a reactor continues
to push new values into the ODB, triggering itself, and never finishing.  A more
subtle issue is that if reactors are "too greedy" and there are lots of writes
going into the ODB, this can cause us to never update, even though our display
is basically complete.

A simple fix for this is to make sure that we don't trigger a second pass
of updating if a reactor doesn't actually do anything.