Cells are 'moveable'.
This means that the same cell can be moved from one part of the tree
to another without having to completely rebuild it. This can be substantially
faster since building large trees of cells can be very slow.

This isn't perfectly worked out - moving cells between contexts where the
parent's scrolling state is different, for example, will probably not work yet.
Similarly, we're not being careful yet to take care of a cell's 'identity path'
which might have changed.