Fix a bug in Editor.js where server-side stream compression could put the client in a bad state.
The server wants to compress and simplify the document event stream so that it doesn't get too
big. When it does so, it may remove events that the front-end has in its buffer, which can cause
it to be able to produce invalid/nonsensical undo events. This change prevents that by
forcing the editor to forget about events before the compression point.